+++ Genesis/src/utopia/genesis/shape/shape2D/TwoDimensional.scala
@@ -1,6 +1,7 @@
 package utopia.genesis.shape.shape2D
 
 import utopia.genesis.shape.Axis.{X, Y}
+import utopia.genesis.shape.Axis2D
 import utopia.genesis.shape.template.Dimensional
 
 /**
@@ -26,4 +27,18 @@ trait TwoDimensional[+A] extends Dimensional[A]
 	  * @return The first two dimensions of this instance
 	  */
 	def dimensions2D = dimensions.take(2)
+	
+	/**
+	  * @return A map that contains this item's first two dimensions (if available), tied to X and Y axes.
+	  */
+	def toMap2D: Map[Axis2D with Product, A] =
+	{
+		val dims = dimensions2D
+		if (dims.size >= 2)
+			Map(X -> dims.head, Y -> dims(1))
+		else if (dims.nonEmpty)
+			Map(X -> dims.head)
+		else
+			Map()
+	}
 }