+++ Reflection/src/utopia/reflection/component/drawing/immutable/RoundedBackgroundDrawer.scala
@@ -1,9 +1,8 @@
 package utopia.reflection.component.drawing.immutable
 
 import utopia.genesis.color.Color
-import utopia.genesis.shape.shape2D.Bounds
-import utopia.genesis.util.Drawer
-import utopia.reflection.component.drawing.template.{CustomDrawer, DrawLevel}
+import utopia.reflection.component.drawing.template
+import utopia.reflection.component.drawing.template.DrawLevel
 import utopia.reflection.component.drawing.template.DrawLevel.Background
 
 object RoundedBackgroundDrawer
@@ -33,23 +32,7 @@ object RoundedBackgroundDrawer
   * @author Mikko Hilpinen
   * @since 12.9.2020, v1.3
   */
-case class RoundedBackgroundDrawer private(color: Color, rounding: Either[Double, Double],
-										   override val drawLevel: DrawLevel = Background) extends CustomDrawer
-{
-	override def draw(drawer: Drawer, bounds: Bounds) =
-	{
-		if (bounds.size.isPositive)
-		{
-			drawer.onlyFill(color).disposeAfter { d =>
-				rounding match
-				{
-					case Left(radius) =>
-						// Won't scale the radius over maximum (circular shape)
-						val maxRadius = bounds.minDimension / 2.0
-						d.draw(bounds.toRoundedRectangleWithRadius(radius min maxRadius))
-					case Right(factor) => d.draw(bounds.toRoundedRectangle(factor))
-				}
-			}
-		}
-	}
-}
+case class RoundedBackgroundDrawer private(override val color: Color,
+                                           override protected val rounding: Either[Double, Double],
+										   override val drawLevel: DrawLevel = Background)
+	extends template.RoundedBackgroundDrawer