+++ Genesis/Changes.md
@@ -17,6 +17,10 @@ new mutable mouse event handlers.
 - Added .oneTimeListener(Filter)(...) to KeyStateListener
 - Added .withOverlay(Image, Point) to Image (can also be called with +)
 - Added new variations of .average(...) to RGB and added copies of those methods to Color
+- Added a new variation of .symmetric(...) to InsetsFactory
+- Added .toMap2D to TwoDimensional
+### Fixes
+- Dimensional.toMap now properly returns a Map
 ### Other Changes
 - Bounds and Point .toAwt conversion now rounds the double values to the closest integers
 - Direction1D now extends RichComparable